fix(circular-dependencies): 解决循环引用，在 ci 中测试循环引用
