+++ package.json
@@ -16,6 +16,7 @@
     "build_test": "rm -rf spec-js && tsc test/unit/app.ts -m commonjs --sourcemap --inlineSources --outDir spec-js --target ES2015 --diagnostics --pretty --experimentalDecorators --suppressImplicitAnyIndexErrors --types \"node,mocha,chai,sinon,sinon-chai,whatwg-fetch\" --moduleResolution node",
     "build_e2e": "rm -rf .tmp/e2e && tsc ./test/e2e/main.ts  -m es2015 --outDir .tmp/e2e --sourcemap --inlineSources  --target ES5 --diagnostics --pretty --experimentalDecorators --suppressImplicitAnyIndexErrors --moduleResolution node && ts-node ./tools/tasks/bundle.e2e.ts",
     "build_socket": "rm -rf dist/bundle/tbsdk.socket.js && tsc ./src/SocketApp.ts -m commonjs --outDir dist/socket --sourcemap --inlineSources --target ES5 -d --diagnostics --pretty --experimentalDecorators --suppressImplicitAnyIndexErrors --moduleResolution  node && ts-node ./tools/tasks/bundle.socket.ts",
+    "check_circular_dependencies": "madge ./dist/cjs --circular",
     "watch": "npm run watch_test & ts-node ./tools/tasks/test.ts",
     "watch_test": "tsc test/unit/app.ts -m commonjs --sourcemap --inlineSources --outDir spec-js --target ES2015 --diagnostics --pretty --experimentalDecorators --suppressImplicitAnyIndexErrors --types \"node,mocha,chai,sinon,sinon-chai,whatwg-fetch\" --moduleResolution node --skipLibCheck --isolatedModules -w",
     "build_bundle": "rm -rf dist/bundle/tbsdk.umd.js && ts-node ./tools/tasks/bundle.sdk.ts",
@@ -73,6 +74,7 @@
     "fetch-mock": "^5.5.0",
     "google-closure-compiler-js": "^20161024.0.0",
     "jsonrpc-lite": "^1.2.1",
+    "madge": "^1.5.0",
     "moment": "^2.16.0",
     "node-watch": "^0.4.1",
     "nyc": "^10.0.0",
