+++ src/schemas/schema.ts
@@ -1,5 +1,5 @@
 'use strict'
-import { forEach, dropEle } from '../utils/index'
+import { forEach, dropEle } from '../utils/helper'
 
 export const bloodyParentMap = new Map<string, string[]>()
